store: move the session actions into their own package
